+++ tests/unit/TmBlock.spec.js
@@ -8,7 +8,7 @@ describe("TmBlock.vue", () => {
     wrapper = mount(TmBlock, {
       propsData: {
         txs,
-        address: "tb1da6xsetjg9jxgun9wdesexv05j",
+        currentUser: "tb1da6xsetjg9jxgun9wdesexv05j",
         candidates: [
           {
             owner: "cosmosvaladdr15ky9du8a2wlstz6fpx3p4mqpjyrm5ctqzh8yqw",