+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Fixed `TmModalError` having redundant button code @nylira
+- Mutate props during test bug @okwme
 
 ## [0.2.12]