+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.15]
+
+- Min and Max to TmField @okwme
+
 ## [0.2.14]
 
 ### Added
 
 - Component previews with `yarn serve`
-- Min and Max to TmField @okwme
 
 ## [0.2.13]